Move some stuff to transients, and small refactors
